feat: sviluppi per prenotazione posti, qualche refactor per maggiore ordine e qualche fix

+++ src/main/java/it/unibo/samplejavafx/cinema/adapter/cliente/ClienteController.java
@@ -18,7 +18,7 @@ public class ClienteController {
   private final ClienteService clienteService;
 
   @GetMapping
-  public Cliente findCliente(@RequestParam int id) {
+  public Cliente findCliente(@RequestParam Long id) {
     try {
       return clienteService.findClienteById(id);
     } catch (Exception e) {
